% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.angdia}
\alias{emov.angdia}
\title{Angular size of stimulus.}
\usage{
emov.angdia(stimsize, distance)
}
\arguments{
\item{stimsize}{Size of the stimulus.}

\item{distance}{Viewing distance from stimulus.}
}
\value{
Angular size in degrees.
}
\description{
Angular size of stimulus.
}

