% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.filter}
\alias{emov.filter}
\title{Velocity threshold filter.}
\usage{
emov.filter(x, y, threshold)
}
\arguments{
\item{x}{Eye position.}

\item{y}{Eye position.}

\item{threshold}{Velocity threshold.}
}
\value{
Filtered data.
}
\description{
Velocity threshold filter.
}

