% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.idt}
\alias{emov.idt}
\title{I-DT algorithm.}
\usage{
emov.idt(t, x, y, dispersion, duration)
}
\arguments{
\item{t}{Vector of timepoints.}

\item{x}{horizontal eye position.}

\item{y}{vertical eye position.}

\item{dispersion}{Maximal dispersion allowed.}

\item{duration}{Minimal fixation duration allowed.}
}
\value{
Fixations: position, start, end.
}
\description{
I-DT algorithm.
}
\examples{
max_disp  = 19.0 # in cm, 28.8 cm (2 deg)
min_dur =  80/1000*200 # 80ms / 200Hz
fix = emov.idt(data$t, data$x, data$y, max_disp, min_dur)
}
\references{
Salvucci, D. D., & Goldberg, J. H. (2000).
Identifying fixations and saccades in eye-tracking protocols. In Proceedings
of the 2000 symposium on eye tracking research & applications (pp. 71-78).
New York: ACM.
}

