% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.read_iviewsamples}
\alias{emov.read_iviewsamples}
\title{Read SMI iview sample file.}
\usage{
emov.read_iviewsamples(file, nr_of_headerlines)
}
\arguments{
\item{file}{Filename.}

\item{nr_of_headerlines}{No. of header lines in datafile.}
}
\value{
data file.
}
\description{
Read SMI iview sample file.
}

