% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.cart2sphere}
\alias{emov.cart2sphere}
\title{Convert Cartesian to Spherical coordinates.}
\usage{
emov.cart2sphere(x, y, z)
}
\arguments{
\item{x}{x.}

\item{y}{y.}

\item{z}{z.}
}
\value{
Two angles and radius-
}
\description{
Convert Cartesian to Spherical coordinates.
}
\examples{
data = emov.cart2sphere(data$L.GVEC.X, data$L.GVEC.Y, data$L.GVEC.Z)
}

