% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emov.R
\name{emov.idt}
\alias{emov.idt}
\title{I-DT algorithm.}
\usage{
emov.idt(t, x, y, dispersion, duration)
}
\arguments{
\item{t}{Vector of timepoints.}

\item{x}{horizontal eye positions.}

\item{y}{vertical eye positions.}

\item{dispersion}{Maximal dispersion allowed (in units of x and y).}

\item{duration}{Minimal fixation duration allowed (in number of samples)}
}
\value{
Fixations: position, start, end.
}
\description{
I-DT algorithm.
}
\references{
Salvucci, D. D., & Goldberg, J. H. (2000).
Identifying fixations and saccades in eye-tracking protocols. In Proceedings
of the 2000 symposium on eye tracking research & applications (pp. 71-78).
New York: ACM.
}

